refactor: split logic and business modules
