+++ business/__init__.py
@@ -1,0 +1,5 @@
+"""Бизнес-логика приложения."""
+
+from .processor import ExcelToGoogleSheets
+
+__all__ = ["ExcelToGoogleSheets"]
