+++ logic/__init__.py
@@ -1,0 +1,9 @@
+"""Низкоуровневые утилиты для работы с таблицами."""
+
+from .sheet_utils import resolve_excel_columns, resolve_google_columns, copy_sheet_data
+
+__all__ = [
+    "resolve_excel_columns",
+    "resolve_google_columns",
+    "copy_sheet_data",
+]
